+++ internal/controller/operator/factory/k8stools/test_helpers.go
@@ -39,7 +39,9 @@ func testGetScheme() *runtime.Scheme {
 		&vmv1beta1.VMClusterList{},
 		&vmv1beta1.VLogsList{},
 		&vmv1.VLSingleList{},
+		&vmv1.VTSingleList{},
 		&vmv1.VLClusterList{},
+		&vmv1.VTClusterList{},
 		&vmv1.VMAnomalyList{},
 		&vmv1.VLAgentList{},
 	)
@@ -63,6 +65,8 @@ func testGetScheme() *runtime.Scheme {
 		&vmv1beta1.VLogs{},
 		&vmv1.VLSingle{},
 		&vmv1.VLCluster{},
+		&vmv1.VTSingle{},
+		&vmv1.VTCluster{},
 		&vmv1.VMAnomaly{},
 		&vmv1.VLAgent{},
 	)
@@ -100,6 +104,8 @@ func GetTestClientWithClientObjects(predefinedObjects []client.Object) client.Cl
 			&vmv1beta1.VMNodeScrape{},
 			&vmv1.VLSingle{},
 			&vmv1.VLCluster{},
+			&vmv1.VTSingle{},
+			&vmv1.VTCluster{},
 			&vmv1.VMAnomaly{},
 			&vmv1.VLAgent{},
 		).