+++ internal/controller/operator/factory/vmauth/vmusers_config.go
@@ -327,6 +327,10 @@ var crdNameToObject = map[string]objectWithURL{
 	"VLCluster/vlinsert":  newClusterWithURL("vlinsert"),
 	"VLCluster/vlstorage": newClusterWithURL("vlstorage"),
 	"VLAgent":             &vmv1.VLAgent{},
+	"VTSingle":            &vmv1.VTSingle{},
+	"VTCluster/vtselect":  newClusterWithURL("vtselect"),
+	"VTCluster/vtinsert":  newClusterWithURL("vtinsert"),
+	"VTCluster/vtstorage": newClusterWithURL("vtstorage"),
 }
 
 // helper interface to restore VMCluster type
@@ -336,13 +340,13 @@ type unwrapObject interface {
 
 var clusterComponentToURL = map[string]func(obj client.Object) string{
 	"vminsert": func(obj client.Object) string {
-		return obj.(*vmv1beta1.VMCluster).VMInsertURL()
+		return obj.(*vmv1beta1.VMCluster).InsertURL()
 	},
 	"vmselect": func(obj client.Object) string {
-		return obj.(*vmv1beta1.VMCluster).VMSelectURL()
+		return obj.(*vmv1beta1.VMCluster).SelectURL()
 	},
 	"vmstorage": func(obj client.Object) string {
-		return obj.(*vmv1beta1.VMCluster).VMStorageURL()
+		return obj.(*vmv1beta1.VMCluster).StorageURL()
 	},
 	"vlinsert": func(obj client.Object) string {
 		return obj.(*vmv1.VLCluster).InsertURL()
@@ -352,6 +356,15 @@ var clusterComponentToURL = map[string]func(obj client.Object) string{
 	},
 	"vlstorage": func(obj client.Object) string {
 		return obj.(*vmv1.VLCluster).StorageURL()
+	},
+	"vtinsert": func(obj client.Object) string {
+		return obj.(*vmv1.VTCluster).InsertURL()
+	},
+	"vtselect": func(obj client.Object) string {
+		return obj.(*vmv1.VTCluster).SelectURL()
+	},
+	"vtstorage": func(obj client.Object) string {
+		return obj.(*vmv1.VTCluster).StorageURL()
 	},
 }
 
@@ -368,6 +381,8 @@ func newClusterWithURL(component string) *clusterWithURL {
 		clusterObj = &vmv1beta1.VMCluster{}
 	case strings.HasPrefix(component, "vl"):
 		clusterObj = &vmv1.VLCluster{}
+	case strings.HasPrefix(component, "vt"):
+		clusterObj = &vmv1.VTCluster{}
 	default:
 		panic(fmt.Sprintf("BUG: unexpected component name: %q", component))
 	}
@@ -724,6 +739,29 @@ func genURLMaps(userName string, refs []vmv1beta1.TargetRef, result yaml.MapSlic
 
 	}
 
+	appendPaths := func(paths []string, tr *vmv1beta1.TargetRef) []string {
+		if tr.CRD != nil {
+			switch tr.CRD.Kind {
+			case "VMCluster/vminsert":
+				if len(tr.TargetPathSuffix) == 0 {
+					return append(paths, "/insert/.*")
+				}
+				return addVMInsertPaths(paths)
+			case "VMCluster/vmselect":
+				if len(tr.TargetPathSuffix) == 0 {
+					return append(paths, "/select/.*", "/admin/.*")
+				}
+				return addVMSelectPaths(paths)
+			case "VLCluster/vlinsert", "VTCluster/vtinsert":
+				return append(paths, "/insert/.*")
+			case "VLCluster/vlselect", "VTCluster/vtselect":
+				return append(paths, "/select/.*")
+			}
+		}
+		return append(paths, "/.*")
+
+	}
+
 	for i := range refs {
 		var urlMap yaml.MapSlice
 		ref := refs[i]
@@ -740,22 +778,7 @@ func genURLMaps(userName string, refs []vmv1beta1.TargetRef, result yaml.MapSlic
 		case 0:
 			// special case for
 			// https://github.com/VictoriaMetrics/operator/issues/379
-			switch {
-			case len(refs) > 1 && ref.CRD != nil && ref.CRD.Kind == "VMCluster/vminsert":
-				if len(ref.TargetPathSuffix) == 0 {
-					paths = append(paths, "/insert/.*")
-				} else {
-					paths = addVMInsertPaths(paths)
-				}
-			case len(refs) > 1 && ref.CRD != nil && ref.CRD.Kind == "VMCluster/vmselect":
-				if len(ref.TargetPathSuffix) == 0 {
-					paths = append(paths, "/select/.*", "/admin/.*")
-				} else {
-					paths = addVMSelectPaths(paths)
-				}
-			default:
-				paths = append(paths, "/.*")
-			}
+			paths = appendPaths(paths, &ref)
 
 		case 1:
 			switch paths[0] {